third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/track/track-element/track-selection-task-order-expected.txt

This is a testharness.js-based test.
[FAIL] HTMLTrackElement Text Track Selection Task Order
  assert_equals: Text track selection hasn't started yet. expected "disabled" but got "showing"
Harness: the test ran to completion.
